Правки по ревью: убрала статический FragmentManager, добавила Listener
