+++ Rebbargtxet/app/src/main/java/com/github/lleuad0/rebbargtxet/NoteViewHolder.kt
@@ -1,6 +1,5 @@
 package com.github.lleuad0.rebbargtxet
 
-import android.os.Bundle
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
@@ -8,23 +7,12 @@ import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.RecyclerView
 import java.text.SimpleDateFormat
 
-const val ID_KEY = "ID_KEY"
-
-class NoteViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+class NoteViewHolder(itemView: View, listener: NoteAdapter.Listener) :
+    RecyclerView.ViewHolder(itemView) {
 
     init {
         val cardView = itemView.findViewById<CardView>(R.id.card_view)
-        cardView.setOnClickListener {
-            val fragmentNoteContent = FragmentNoteContent()
-            val bundle = Bundle()
-            bundle.putLong(ID_KEY, id)
-            fragmentNoteContent.arguments = bundle
-
-            MainActivity.manager.beginTransaction()
-                .replace(R.id.fragment_holder, fragmentNoteContent)
-                .addToBackStack(null)
-                .commit()
-        }
+        cardView.setOnClickListener { listener.onNoteClick(id) }
     }
 
     private val previewTextView = itemView.findViewById<TextView>(R.id.preview_text_view)
